==> Project_3/code/1.2_Relationship Advice Scrapping.ipynb <==
## Part 1.2 - Data Scrapping

After scrapping the data for marriage subreddit, I will scrap the data for relationship subreddit.

# Importing required library
import requests
import pandas as pd

### Data Scrapping of relationship subreddit

Recreating the function used previously

# Instantiating the function
def reddit_pushshift(reps, utc_int=None, subreddit, final_df):
    
    # Setting the initial time of the post
    last_utc = utc_int
    
    # For loop to extract the post until the desired number of posts
    for rep in range(reps):
        
        # Setting the base url and params
        url = 'https://api.pushshift.io//reddit/search/submission/'
        params = {
        'subreddit': subreddit,
        'size': 100,
        'before': last_utc
        }
        
        res = requests.get(url, params)
        data = res.json()
        posts = data['data']

        # Create dataframe of posts with desired columns
        df = pd.DataFrame(posts)

        df = df[['selftext', 'title', 'created_utc', 'id', 'score']]
        
        # Concatenate dataframe to an existing dataframe
        final_df = pd.concat(
            objs=[final_df, df],
            axis=0,
        )
        
        # Resetting the index after concatenating
        final_df.reset_index(drop=True, inplace=True)
        
        # Reset the time factor to the time of last post
        last_utc = final_df['created_utc'][-1:]
        
        # To check on the progress of the extraction
        print(rep, last_utc)
    
    return final_df


# Creating an empty dataframe to append the extracted posts.
relationship_df = pd.DataFrame(columns = ['selftext', 'title', 'created_utc', 'id', 'score'])

relationship_df

# Extract 100 post 100 times.
relationship_df100 = reddit_pushshift(100, None, 'relationship',  relationship_df)

# Check on the posts that were extracted
relationship_df100

# Create csv for data cleaning later
relationship_df100.to_csv('relationship_advice_all_posts.csv', index=False)

### Proceed to next notebook for data cleaning.
