==> Project_3/code/2_Data Cleaning Overall.ipynb <==
## Part 2 - Data Cleaning

After extracting the data from reddit, I will clean the datasets.

# Importing required libraries
import pandas as pd
import re

from nltk.stem import WordNetLemmatizer
from nltk.corpus import stopwords
from sklearn.feature_extraction.text import CountVectorizer

# Import datasets which were extracted
marriage = pd.read_csv('../datasets/Marriage_Advice_all_posts.csv')
relationship = pd.read_csv('../datasets/Relationship_all_posts.csv')

### Cleaning marriage dataset

marriage

# Fill nan with an empty space instead.
marriage = marriage.fillna('')

marriage

# Dropping score and id column as they will not be used
marriage.drop(axis=1, columns=['score', 'id'], inplace=True)

There is a need to remove posts which were deleted or removed by moderators. There is also a need to remove repeated posts by moderators to remind redditors on the rules of that subreddit.

# Creating a mask to remove removed, deleted and moderator posts.
to_remove = (
    ((marriage['selftext'] == '[removed]')) |
    ((marriage['selftext'] == '[deleted]'))| 
    (marriage['title'].str.contains('Weekly advice')) 
)

# Drop rows which were identified
marriage.drop(marriage[to_remove].index, inplace=True)

# Drop any other duplicated posts
marriage.drop_duplicates(inplace=True)

While cleaning the data, I have noticed that there are some other languages being used in this thread which will hinder the analysis. Hence I will remove non-ascii characters as well as html_markers.

marriage['title'] = marriage['title'].str.encode('ascii', 'ignore').str.decode('ascii')

marriage['selftext'] = marriage['selftext'].str.encode('ascii', 'ignore').str.decode('ascii')

# Creating a function to strip some identified html markers
def remove_html_markers(text):
    
    text = re.sub('xa0', '', str(text)).strip()
    text = re.sub('x200b', '', str(text)).strip()
    
    return text

# Applying function to the data sets
marriage['selftext'] = marriage['selftext'].map(remove_html_markers)
marriage['title'] = marriage['title'].map(remove_html_markers)

Creating extra features for further analysis

# Creating a columns made from the combined text of title and selftext
marriage['all_text'] = marriage['title'] + " " + marriage['selftext']

# Creating the columns to get the characther length of each posts
marriage['text_length'] = marriage['selftext'].map(lambda x: len(x))
marriage['title_length'] = marriage['title'].map(lambda x: len(x))
marriage['all_text_length'] = marriage['all_text'].map(lambda x: len(x))

# Labelling marriage as the target post
marriage['is_marriage'] = 1

# Reset index 
marriage.reset_index(drop=True, inplace=True)

marriage

After cleaning up the dataset, there are still 9313 posts.

### Cleaning relationship dataset

relationship

# Fill nan with an empty space instead.
relationship = relationship.fillna('')

relationship

# Dropping score and id column as they will not be used
relationship.drop(axis=1, columns=['score', 'id'], inplace=True)

There is a need to remove posts which were deleted or removed by moderators. I have also identified a post which was was a spam post and hence i will remove it as well.

# Creating a mask to remove removed, deleted and spam posts.
to_remove_relationship = (
    (relationship['selftext'] == '[removed]') | 
    (relationship['selftext'] == '[deleted]') |
    (relationship['selftext'].str.contains('lower surface'))
)

# Drop rows which were identified
relationship.drop(relationship[to_remove_relationship].index, inplace=True)

# Drop any other duplicated posts
relationship.drop_duplicates(inplace=True)

# Creating a function to strip some identified html markers
relationship['selftext'] = relationship['selftext'].map(remove_html_markers)
relationship['title'] = relationship['title'].map(remove_html_markers)

Creating extra features for further analysis

# Creating a columns made from the combined text of title and selftext
relationship['all_text'] = relationship['title'] + " " + relationship['selftext']

# Creating the columns to get the characther length of each posts
relationship['text_length'] = relationship['selftext'].map(lambda x: len(x))
relationship['title_length'] = relationship['title'].map(lambda x: len(x))
relationship['all_text_length'] = relationship['all_text'].map(lambda x: len(x))

# Labelling relationship as 0
relationship['is_marriage'] = 0

# Reset index 
relationship.reset_index(drop=True, inplace=True)

relationship

After cleaning up the dataset, there are still 8841 posts left.

### Add lemmatize column

Creating a column to lemmatize the words to remove some of the repeated words. This may help to increase the model accuracy as words such as {'he', 'she', 'in', 'and', 'to' and 'a'} does not help in identifying the models.

# Creating a stopword set to be used to lemmatize
stop_words1 = set(CountVectorizer(stop_words = 'english').get_stop_words())

def lemmatize(text):
    # Split each word for lemmatization
    words = text.split(' ')
    
    # Lemmatize words.
    lemmatizer = WordNetLemmatizer()
    remaining_words = [lemmatizer.lemmatize(w) for w in words if not w in stop_words1]
    
    return(" ".join(remaining_words))

# Creating a column for lemmatized words
marriage['lem_all_text'] = marriage['all_text'].apply(lemmatize)
relationship['lem_all_text'] = relationship['all_text'].apply(lemmatize)

### Combining the datasets

After removing unwanted characters and posts, as well as creating various columns for further analysis, I will now combine both datasets so that they may be used for modelling.

# Combine the two cleaned datasets
combined = pd.concat(
    objs=[relationship, marriage],
    axis=0,
)

# Reset index
combined.reset_index(drop=True, inplace=True)

# Check to ensure that the datasets are concatenated properly
combined

Save the cleaned datasets as CSV

relationship.to_csv('../datasets/relationship_cleaned.csv', index=False)

marriage.to_csv('../datasets/marriage_cleaned.csv', index=False)

combined.to_csv('../datasets/combined.csv', index=False)

### Proceed to next notebook for EDA.
