==> Project_3/code/5_Final Model and Evaluation.ipynb <==
## Part 5 Final Model and Evaluation

After choosing the final model, I will evaluate the model and perform error analysis on the model.

# Importing required libaries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.pipeline import Pipeline
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import (
    confusion_matrix, 
    plot_confusion_matrix,
    accuracy_score,
    plot_roc_curve,
    roc_auc_score,
    recall_score,
    precision_score,
    f1_score,
    RocCurveDisplay
)   


from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer

**Summary Table**

|Model|Vectorizer| Train score | Test Score | ROC AUC Score| TN | FP | FN | TP|
|---|---|---|---|---|---|---|---|---|
|Benchmark Model|CVEC|89.3%|83.4%| 90.3% |1446|322|280|1583|
|MultinomialNB |CVEC|86.4%|83.6%|90.2%|1460|308|288|1575|
|MultinomialNB |TVEC|86.8%|83.4%|91.3%|1437|331|271|1592|
|<font color='red'>LogisticRegression</font>|<font color='red'>CVEC</font>|94.7%|84.7%|91.8%|1466|302|255|1608|
|LogisticRegression|TVEC|91.0%|85.1%|92.7%|1477|291|252|1611|
|RandomForest|CVEC|100%|83.4%|91.9%|1460|308|294|1569|
|RandomForest|TVEC|99.99%|82.9%|91.4%|1453|315|305|1558|


# Importing datasets
marriage = pd.read_csv('../datasets/marriage_cleaned.csv')
relationship = pd.read_csv('../datasets/relationship_cleaned.csv')
combined = pd.read_csv('../datasets/combined.csv')
custom_stop = pd.read_csv('../datasets/custom_stop_words.csv')

# Instantiating the custom_stop_words created previously
custom_stop_words = custom_stop['words'].tolist()

X = combined['all_text']
y = combined['is_marriage']

# Creating a train test split for modelling
X_train, X_test, y_train, y_test = train_test_split(X,
                                                    y,
                                                    test_size=0.2,
                                                    stratify=y,
                                                    random_state=42)

# Creating a pipeline with previously identified best params
pipe_final_model = Pipeline([
    ('cvec', CountVectorizer(
        max_df=0.95,
        max_features=6000,
        min_df=2,
        ngram_range=(1, 2),
        stop_words = custom_stop_words
    )),
    ('logreg', LogisticRegression(
        C=0.1,
        max_iter=1000,
        penalty='l1',
        solver='liblinear'
    ))],
    verbose=2
)



# Fitting the model with the training data.
pipe_final_model.fit(X_train, y_train)

# Scoring the training data
pipe_final_model.score(X_train, y_train)

# Scoring the test data
pipe_final_model.score(X_test, y_test)

### Error Analysis

# Creating a dataframe of predictions on the combined dataset
predictions = pd.DataFrame(pipe_final_model.predict_proba(X))

# Changing the columns to relationship and marriage
predictions.columns = ['Relationship', 'Marriage']

# Adding the original posts and actual classification
predictions['all_text'] = combined['all_text']
predictions['Actual'] = combined['is_marriage']

# Adding the predicted classification
predictions['Predicted'] = pipe_final_model.predict(X)

# Rounding the probability to 2 decimal places
predictions['Relationship'] = predictions['Relationship'].map(lambda x: round(x, 2))
predictions['Marriage'] = predictions['Marriage'].map(lambda x: round(x, 2))

predictions

# Reordering the columns
predictions = predictions.loc[:,[
    'all_text',
    'Actual',
    'Predicted',
    'Relationship',
    'Marriage'
]]

predictions

# All the False Positives as well as False Negatives
wrongly_predicted = predictions[predictions['Actual'] != predictions['Predicted']]

wrongly_predicted 

# False Positive
marriage_in_relationship = wrongly_predicted[
    ((wrongly_predicted['Actual'] == 0 ) & (wrongly_predicted['Predicted'] == 1))]


marriage_in_relationship.head(10)

One particular problem noticed is that relationship posts contains posts made by marriage couples. This may cause confusion the model and the impact is required to be further investigated.

# Creating a mask to identify married couples
married = ((marriage_in_relationship['all_text'].str.contains('husband')) | 
                ((marriage_in_relationship['all_text'].str.contains('wife')))|
                ((marriage_in_relationship['all_text'].str.contains('spouse'))) |
                (marriage_in_relationship['all_text'].str.contains('marr')))

# Checking number of married couples in wrongly classified relationship. 
marriage_in_relationship.loc[married]

Only approximately 400 married couples were posting wrongly in the relationship subreddit. This amounts to approximately 4.5% of the overall relationship subreddit which is an acceptable margin of error. Hence, this impact is minimal.

Furthermore, some of the posts were not made by married couples. It could have been written by unmarried couples who are asking advice about marriage or their future husband or wives. 

An example can be seen below.

marriage_in_relationship['all_text'][8761]

# False Negative
relationship_in_marriage = wrongly_predicted[
    ((wrongly_predicted['Actual'] == 1 ) & (wrongly_predicted['Predicted'] == 0))]


relationship_in_marriage.head(10)

Most of the posts seemed to be wrongly misclassified due to the strict stop words used.

### Conclusion and Recommendation

Using LogisticRegression classifier with a CountVectorizer, I was able to achieve an accuracy of 84.6% which is quite reasonable. Being able to differentiate the posts would allow me to identify concerns which are relevant to married couples and unmarried couples.

Next step for me would be to create the FAQ regarding the issues which were identified.

Other areas for future exploration:
- I will be looking to incorporate a sentiment library which was specially tailored to relationship issue and to subsequently perform sentiment analysis on the posts. This will allow for more worrying issues to be more prominent.

- Some of the previously identified investigation points were lost during the modelling and were not able to be investigated. I will specifically look into those points.
